Change the decoding process for FFmpeg 3.1
This introduces a new UncompressedSampleProvider that contains the logic to send frames to be decoded by FFmpeg.
This helps consolidate the code between the Audio and Video uncompressed sample providers.

As packets are being decoded the following is occuring:
Try to feed the packet to the decoder. If the decoder is full, put it back in the queue.
Try to get a decoded frame from the decoder. If the decoder doesn't have any frames, then flag that as incomplete and process the next packet.

When creating a MediaFoundation sample, get the timestamp from the decoded frame instead of the from the packet.